fusor ui - updating to move js and stylesheets to fusor_ui subdirectory
This commit moves the javascript and stylesheet files to a fusor_ui subdirectory.
This makes them consistent with other engines and helps to ensure that during installation
the files will not clobber any foreman core files.

This doesn't move the images; however, we should consider doing those as well.
It was not done with this commit, since the javascript refers to the images
at assets/images.