Add deploy rhev action.
The deploy rhev action will inspect the deployment object and determine if it
needs to deploy a RHEV Engine & Hypervisors. It will find the proper hostgroups
for each, assign them to the proper host. Convert any discovered hosts to
managed hosts for provisioning.

There is quite a bit of DEBUG logging marked with XXX. Normally this wouldn't be
kept but while we're still anticipating some weird bugs it will be helpful until
we can iron them all out. The XXX makes it easier to find the messages in the
log file as well as making it easier to find them for removal later.