fusor server: update config and deploy action to support overriding parameters on hostgroups
This commit contains changes to enable for the deploy action
to provide override values on hostgroup puppet class parameters.
This is initially needed to allow for the ovirt puppet class to have
the product set to 'RHEV'.

This commit also enables the override flag on several additional
smart class parameters for the ovirt puppet classes.