fusor server: support deploying using the 'Default Organization View' content view
The changes provided in this commit will allow a user to perform
a deployment using RPM content that is in the 'Default Organization
View'.  This view is often referred to as the 'default content view'
and is published when the user syncs the repository.

In order to perform a deployment using the 'Default Organization View',
simply set 'deployment.lifecycle_environment_id = nil' prior to
performing the deploy.

When this is done,

- RPM content will come from the Default Organization View

- Puppet content will come from the 'Fusor Puppet Content' view
  currently in the Library.  Puppet content is not part of the
  'Default Organization View'; however, this particular content
  view is created and published as part of the fusor-installer.