fusor server: decouple puppet class parameter configuration from the hostgroups
This commit contains a change that will move the definition of the
puppet classes supported outside of the hostgroup.  In addition,
it adds logic that will enable the parameter override for all parameters
provided by the puppet class (versus requiring them to be listed
in the configuration).

These changes will enable deployments to support the following:
- associating of puppet classes with a host group based upon configuration
- associating of puppet classes with a host or host group outside of
  the configuration - this scenario will be more useful when we
  support the cloud forms scenario which will require performing
  multiple puppet runs