fusor server: re-order behavior of the activation creation action
Previously, we utilized the activation key create action from
Katello.  This worked well; however, we needed to re-order some
of the actions done by deploy so that more work is done in the run
phase vs finalize.   In doing that, we found that the activation
key create finalize needs to occur before we can add subscriptions
in the run phase of another action.

In order to address this, we are adding an action to inherit
from Katello's create and simply have the new action perform
the work of Katello's finalize in the run phase.