fusor server: support 'adding rhev provider' to cloudforms VM
This commit contains the initial changes to support adding
a RHEV provider to the CloudForms VM, using the mechanize
gem.

Note: the call to add the provider is currently commented out,
as we need to have the IP address of the CloudForms VM, when it
is available.  That will be handled in a future commit.