fusor server: update trigger provisioning of hosts to be in sequence
This commit will cause the trigger of provisioning for the hypervisor
hosts to be done in 'sequence' vs using 'concurrence' (i.e. parallel).
This is to address a deadlock issue that was observed by the simultaneous
'host.save!' that occurs when the trigger occurs.

The error observed looked similar to:

ActiveRecord::StatementInvalid: PGError: ERROR:  deadlock detected
DETAIL:  Process 7957 waits for ShareLock on transaction 74849; blocked by process 7961.
Process 7961 waits for ExclusiveLock on tuple (0,7) of relation 18181 of database 17944; blocked by process 7957.
HINT:  See server log for query details.
: UPDATE "environments" SET "hosts_count" = 1 WHERE "environments"."id" = 2