fusor server: update foreman puppet classes during content view publishing
This commit addresses an issue where we observed puppet classes
not getting assigned to the hostgroups when a deployment
was using a lifecycle environment (e.g. Library, dev...etc).
The issue was that the puppet classes were not being imported
in to foreman until the 'finalize' phase (of the the
content view publish/promotion action).  This causes troubles for
fusor, as we need to have those imported during the run phase
so that they are available for the hostgroup assignments.