fusor server: refactor the deploy cloudforms action in to mulitple actions
This commit is a refactor of the current deploy_cloudforms action
which performs all of the steps for deploying cloudforms as a single
action.  While this approach is the simplist, if any error occurs during
execution, the user must re-execute all steps within the action, which
is really not ideal.

This commit will decompose the current mega-action in to several smaller
actions representing each of the steps required.

Note: this commit does not address other issues raised with this
action.  Those will be worked as separate commits.