Adding a sleep between saving a Host to build it and rebooting it.
This is a short term workaround until we can learn a way to determine if the
required callbacks have executed after the Host.save! has completed.

Without this change we run the risk of intermittent problems provisioning a node.
We've seen the problems with 2+ hypervisors, doesn't seem to appear with a single
node being provisioned