remove sleep and hard coded reboot
* use to_managed to convert Discovered -> Managed.
* move host.save outside of assign_host_to_hostgroup
* Also don't delete the virtual interfaces.
* removed warn log statements, changed some to debug
* move save inside assign & within txn

For some reason using save! breaks things. Also doing the save outside of the
assign_host_to_hostgroup method would not save the host to the db.

I switched to using a simple save, and moved ::Host.transaction inside the
assign_* method. I also moved the save into the method as well (it needs to
remain inside the transaction).