log out what assign host returns
this will shut rubocop up, and still allow me to have the variable
assigned in case someone comes in afterwards and wants to use it.
I'm not sure what ruby will do with the host variable that was passed into the
method.

I.e.

  1. host = something
  2. assign_host_to_hostgroup(host, hostgroup)
  3. host.do_something_to_it.

will host on line 3 be the one updated in assign method? or will it be the oen
from line 1?