Disable deploy if no manifest_file is set
tr796: If backing up and choosing disconnected *after* making it
to the review page, and not uploading a manifest, it's possible
to arrive at the review page and initiate a deployemnt without
a manifest. This causes a deployment to blowup.

This patch prevents you from being able to actually deploy without
a manifest and allows you to backup and set one in order to proceed.