Content Sync Failure Recovery
> Upon an error thrown during content sync, ManageContent task will
be auto-resumed to ensure locks are dropped.

> During progress overview, will monitor for ManageContent error.
If this occurs, will mark the deployment with a content error
`has_content_error` in the fusor database and prompt the user
with recovery options.

> Deployment can be abandoned (deleted), or the user can attempt a
redeployment. In practice, this means fusor server kicks off a new
Deploy task using the existing deployment object, and the front
end re-associates the new task with the deployment.