1349559 - Added RegEx to PasswordFilter.rb
Added RegEx to PasswordFilter.rb to ensure that passwords are not filtered from labels.
For example, if the user selects 'password' as their master pass, field label 'rhev_password'
should not show up as 'rhev_[FILTERED]'. Filtering in such a case could allow passwords to be
guessed from reading the log file.