1306823 - hypervisor table shows engine disabled
This commit mirrors a previous commit in which functionality was added
to display all discovered hosts in the engine selection table, but show
selected hypervisors as disabled TRs with tooltips explaining.

Contents of this commit are same as mentioned above, but for the
hypervisor table. The selected RHV Engine will show up in the table as
disabled with a helpful tooltip.