rhev_cpu_type gets the default cpu family detected
Fixed a bug where rhev_cpu_type was only being set on the deployment
object when the CPU family selection box was manually modified.

This could potentially cause a deploy to fail on AMD and IBM machines
if the CPU family selection box was left unmodified by the user. On
Intel machines, the null value for rhev_cpu_type would result in the
overall default (Intel Nehalem) being selected.